+++ android/src/main/java/co/pushe/plus/flutter/PushePlugin.java
@@ -4,25 +4,24 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.util.Log;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+
 import co.pushe.plus.Pushe;
-import co.pushe.plus.notification.PusheNotification;
 import co.pushe.plus.analytics.PusheAnalytics;
+import co.pushe.plus.notification.PusheNotification;
 import co.pushe.plus.notification.UserNotification;
-import io.flutter.plugin.common.JSONMethodCodec;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 import io.flutter.view.FlutterMain;
-import java.util.Map;
-import java.util.List;
-
 
 
 /**
@@ -31,9 +30,11 @@ import java.util.List;
  * @author Mahdi Malvandi
  * FIXME: Warnings will be repaired with the next release
  */
+@SuppressWarnings("NullableProblems")
 public class PushePlugin implements MethodCallHandler {
 
     private Context context;
+    private final List<String> NOTIF_TYPES = Arrays.asList("IdType.AndroidId", "IdType.GoogleAdvertisingId", "IdType.CustomId");
 
     public static void registerWith(Registrar registrar) {
         final MethodChannel channel = new MethodChannel(registrar.messenger(), "Pushe");
@@ -52,9 +53,13 @@ public class PushePlugin implements MethodCallHandler {
         context.registerReceiver(new PusheNotificationReceiver(new MethodChannel(registrar.messenger(), "Pushe")), i);
     }
 
+    // TODO: Fix long method issue
     @Override
     public void onMethodCall(MethodCall call, final Result result) {
         String methodName = call.method;
+        PusheNotification notificationModule = Pushe.getPusheService(PusheNotification.class);
+        PusheAnalytics analyticsModule = Pushe.getPusheService(PusheAnalytics.class);
+
         switch (methodName) {
             case "Pushe.getAndroidId":
                 result.success(Pushe.getAndroidId());
@@ -74,7 +79,7 @@ public class PushePlugin implements MethodCallHandler {
                     Pushe.setCustomId(id);
                     result.success("Will set custom id " + id);
                 } else {
-                    result.error("404", "Failed to set custom id. No id provided.", null);
+                    result.error("001", "Failed to set custom id. No id provided.", null);
                 }
                 break;
 
@@ -88,7 +93,7 @@ public class PushePlugin implements MethodCallHandler {
                     Pushe.setUserEmail(email);
                     result.success("Will set user email to " + email);
                 } else {
-                    result.error("404", "Failed to set email. No email provided.", null);
+                    result.error("002", "Failed to set email. No email provided.", null);
                 }
                 break;
 
@@ -102,7 +107,7 @@ public class PushePlugin implements MethodCallHandler {
                     Pushe.setUserPhoneNumber(phone);
                     result.success("Will set user phone to " + phone);
                 } else {
-                    result.error("404", "Failed to set phone. No phone provided.", null);
+                    result.error("003", "Failed to set phone. No phone provided.", null);
                 }
                 break;
 
@@ -117,13 +122,13 @@ public class PushePlugin implements MethodCallHandler {
                         }
                     });
                 } else {
-                    result.error("404", "Failed to subscribe. No topic argument is passed", null);
+                    result.error("004", "Failed to subscribe. No topic argument is passed", null);
                 }
                 break;
             case "Pushe.unsubscribe":
                 if (call.hasArgument("topic")) {
                     String topic = call.argument("topic");
-                    Pushe.unsubscribeFromTopic( topic, new Pushe.Callback() {
+                    Pushe.unsubscribeFromTopic(topic, new Pushe.Callback() {
                         @Override
                         public void onComplete() {
                             // Done
@@ -132,17 +137,29 @@ public class PushePlugin implements MethodCallHandler {
                     });
 
                 } else {
-                    result.error("404", "Failed to unsubscribe. No topic provided.", null);
+                    result.error("005", "Failed to unsubscribe. No topic provided.", null);
                 }
                 break;
             case "Pushe.disableNotifications":
-                Pushe.getPusheService(PusheNotification.class).disableNotifications();
+                if (notificationModule == null) {
+                    result.error("013", "Failed to interact with notification module", null);
+                    return;
+                }
+                notificationModule.disableNotifications();
                 break;
             case "Pushe.enableNotifications":
-                Pushe.getPusheService(PusheNotification.class).enableNotifications();
+                if (notificationModule == null) {
+                    result.error("014", "Failed to interact with notification module", null);
+                    return;
+                }
+                notificationModule.enableNotifications();
                 break;
             case "Pushe.isNotificationEnable":
-                result.success(Pushe.getPusheService(PusheNotification.class).isNotificationEnable());
+                if (notificationModule == null) {
+                    result.error("015", "Failed to interact with notification module", null);
+                    return;
+                }
+                result.success(notificationModule.isNotificationEnable());
                 break;
             case "Pushe.isInitialized":
                 result.success(Pushe.isInitialized());
@@ -150,34 +167,144 @@ public class PushePlugin implements MethodCallHandler {
             case "Pushe.isRegistered":
                 result.success(Pushe.isRegistered());
                 break;
-            case "Pushe.sendNotificationToUser":
-                if (call.hasArgument("androidId")
-                        && call.hasArgument("title")
-                        && call.hasArgument("content")) {
-                    UserNotification userNotification = UserNotification.withAndroidId((String) call.argument("androidId"));
-                    userNotification.setTitle((String) call.argument("title"));
-                    userNotification.setContent((String) call.argument("content"));
-                    Pushe.getPusheService(PusheNotification.class).sendNotificationToUser(userNotification);
+
+            case "Pushe.sendUserNotification":
+                //noinspection RedundantCast
+                if (call.hasArgument("type") && NOTIF_TYPES.contains((String) call.argument("type")) &&
+                        call.hasArgument("id") &&
+                        call.hasArgument("title") && call.hasArgument("content")) {
+                    String type = call.argument("type");
+                    String id = call.argument("id");
+                    if (id == null || type == null) {
+                        result.error("006", "The keys type and id must exist", null);
+                        return;
+                    }
+                    if (notificationModule == null) {
+                        result.error("007", "Failed to interact with notification module", null);
+                        return;
+                    }
+                    UserNotification notification = null;
+                    switch (type) {
+                        case "IdType.AndroidId":
+                            notification = UserNotification.withAndroidId(id);
+                            break;
+                        case "IdType.GoogleAdvertisingId":
+                            notification = UserNotification.withAdvertisementId(id);
+                            break;
+                        case "IdType.CustomId":
+                            notification = UserNotification.withCustomId(id);
+                            break;
+                        default:
+                            break;
+                    }
+
+                    if (notification == null) {
+                        result.error("008", "Type must be either 'AndroidId', 'GoogleAdvertisingId' or 'CustomId'", null);
+                        return;
+                    }
+
+                    String title = call.argument("title");
+                    String content = call.argument("content");
+                    String bigTitle = call.argument("bigTitle");
+                    String bigContent = call.argument("bigContent");
+                    String imageUrl = call.argument("imageUrl");
+                    String iconUrl = call.argument("iconUrl");
+                    String notifIcon = call.argument("notifIcon");
+                    String customContent = call.argument("customContent");
+
+                    notification.setTitle(title)
+                            .setContent(content)
+                            .setBigTitle(bigTitle)
+                            .setBigContent(bigContent)
+                            .setImageUrl(imageUrl)
+                            .setIconUrl(iconUrl)
+                            .setNotifIcon(notifIcon)
+                            .setCustomContent(customContent);
+
+                    notificationModule.sendNotificationToUser(notification);
+
+                } else {
+                    result.error("009", "Failed to send notification", null); // TODO: better crash
                 }
                 break;
+            case "Pushe.sendAdvancedUserNotification":
+                //noinspection RedundantCast
+                if (call.hasArgument("type") && NOTIF_TYPES.contains((String) call.argument("type")) &&
+                        call.hasArgument("id") &&
+                        call.hasArgument("advancedJson")) {
+                    String type = call.argument("type");
+                    String id = call.argument("id");
+                    if (id == null || type == null) {
+                        result.error("020", "The keys 'type' and 'id' must exist", null);
+                        return;
+                    }
+                    if (notificationModule == null) {
+                        result.error("021", "Failed to interact with notification module", null);
+                        return;
+                    }
+                    UserNotification notification = null;
+                    switch (type) {
+                        case "IdType.AndroidId":
+                            notification = UserNotification.withAndroidId(id);
+                            break;
+                        case "IdType.GoogleAdvertisingId":
+                            notification = UserNotification.withAdvertisementId(id);
+                            break;
+                        case "IdType.CustomId":
+                            notification = UserNotification.withCustomId(id);
+                            break;
+                        default:
+                            break;
+                    }
 
+                    if (notification == null) {
+                        result.error("008", "Type must be either 'AndroidId', 'GoogleAdvertisingId' or 'CustomId'", null);
+                        return;
+                    }
+
+                    String advancedJson = call.argument("advancedJson");
+
+                    notification.setAdvancedNotification(advancedJson);
+
+                    notificationModule.sendNotificationToUser(notification);
+
+                } else {
+                    result.error("009", "Failed to send notification", null); // TODO: better crash
+                }
+                break;
             case "Pushe.sendEvent":
                 if (call.hasArgument("name")) {
                     String name = call.argument("name");
-                    Pushe.getPusheService(PusheAnalytics.class).sendEvent(name);
+                    if (name == null) {
+                        result.error("016", "'name' can not be null", null);
+                        return;
+                    }
+                    if (analyticsModule == null) {
+                        result.error("017", "Failed to interact with analytics module", null);
+                        return;
+                    }
+                    analyticsModule.sendEvent(name);
                     result.success("Will send event " + name);
                 } else {
-                    result.error("404", "Failed to send event. No event name provided.", null);
+                    result.error("010", "Failed to send event. No event name provided.", null);
                 }
                 break;
             case "Pushe.sendEcommerceData":
                 if (call.hasArgument("name") && call.hasArgument("price")) {
                     String name = call.argument("name");
                     Double price = call.argument("price");
-                    Pushe.getPusheService(PusheAnalytics.class).sendEcommerceData(name,price);
+                    if (name == null || price == null) {
+                        result.error("019", "'name' and 'price' can not be null", null);
+                        return;
+                    }
+                    if (analyticsModule == null) {
+                        result.error("018", "Failed to interact with analytics module", null);
+                        return;
+                    }
+                    analyticsModule.sendEcommerceData(name, price);
                     result.success("Will send ecommerce data " + name);
                 } else {
-                    result.error("404", "Failed to send ecommerce data. No event name and price provided.", null);
+                    result.error("011", "Failed to send ecommerce data. No event name and price provided.", null);
                 }
 
                 break;
@@ -193,7 +320,7 @@ public class PushePlugin implements MethodCallHandler {
                         // Done
                         result.success(true);
                     }
-                } );
+                });
                 break;
 
             case "Pushe.setInitializationCompleteListener":
@@ -203,13 +330,13 @@ public class PushePlugin implements MethodCallHandler {
                         // Done
                         result.success(true);
                     }
-                } );
+                });
                 break;
 
             case "Pushe.addTags":
                 if (call.hasArgument("tags") && call.argument("tags") instanceof Map) {
-                    Map tags = call.argument("tags");
-                    Pushe.addTags( tags, new Pushe.Callback() {
+                    Map<String, String> tags = call.argument("tags");
+                    Pushe.addTags(tags, new Pushe.Callback() {
                         @Override
                         public void onComplete() {
                             // Done
@@ -218,14 +345,14 @@ public class PushePlugin implements MethodCallHandler {
                     });
 
                 } else {
-                    result.error("404", "Failed to add tags. No tags provided.", null);
+                    result.error("012", "Failed to add tags. No tags provided.", null);
                 }
                 break;
 
             case "Pushe.removeTags":
                 if (call.hasArgument("tags") && call.argument("tags") instanceof List) {
-                    List tags = call.argument("tags");
-                    Pushe.removeTags( tags, new Pushe.Callback() {
+                    List<String> tags = call.argument("tags");
+                    Pushe.removeTags(tags, new Pushe.Callback() {
                         @Override
                         public void onComplete() {
                             // Done
@@ -234,7 +361,7 @@ public class PushePlugin implements MethodCallHandler {
                     });
 
                 } else {
-                    result.error("404", "Failed to remove tags. No tags provided.", null);
+                    result.error("012", "Failed to remove tags. No tags provided.", null);
                 }
                 break;
 
@@ -279,14 +406,7 @@ public class PushePlugin implements MethodCallHandler {
             } else if (action.equals(context.getPackageName() + ".nc")) {
                 channel.invokeMethod("Pushe.onNotificationClicked", getNotification(intent).toString());
             } else if (action.equals(context.getPackageName() + ".nbc")) {
-                try {
-                    JSONObject o = new JSONObject();
-                    o.put("notification", getNotification(intent).toString());
-                    o.put("button", getButton(intent).toString());
-                    channel.invokeMethod("Pushe.onNotificationButtonClicked", o);
-                } catch (JSONException e) {
-                    Log.w("Pushe", "Failed to parse json", e);
-                }
+                channel.invokeMethod("Pushe.onNotificationButtonClicked", getNotificationAndButton(intent).toString());
             } else if (action.equals(context.getPackageName() + ".nccr")) {
                 channel.invokeMethod("Pushe.onCustomContentReceived", getJson(intent).toString());
             } else if (action.equals(context.getPackageName() + ".nd")) {
@@ -320,9 +440,9 @@ public class PushePlugin implements MethodCallHandler {
             JSONObject o = new JSONObject();
             try {
                 String button = intent.getStringExtra("button");
-                o.put("buttons", new JSONObject(button));
+                o = new JSONObject(button);
             } catch (JSONException e) {
-               System.out.println("Pushe: Failed to parse notification button");
+                System.out.println("Pushe: Failed to parse notification button");
             }
             return o;
         }